go version works
