+++ spec/spec_helper.rb
@@ -7,17 +7,17 @@ Bundler.setup
 require "ethon"
 require 'rspec'
 
-require_relative 'support/localhost_server'
-require_relative 'support/server'
-
-# Ethon.logger = Logger.new($stdout).tap do |log|
-#   log.level = Logger::DEBUG
+# require_relative 'support/localhost_server'
+# require_relative 'support/server'
+#
+# # Ethon.logger = Logger.new($stdout).tap do |log|
+# #   log.level = Logger::DEBUG
+# # end
+#
+# RSpec.configure do |config|
+#   # config.order = :rand
+#
+#   config.before(:suite) do
+#     LocalhostServer.new(TESTSERVER.new, 3001)
+#   end
 # end
-
-RSpec.configure do |config|
-  # config.order = :rand
-
-  config.before(:suite) do
-    LocalhostServer.new(TESTSERVER.new, 3001)
-  end
-end
